arch: fix ports::store layer violation — move ModelDef to domain (#170)
Move ModelDef and TransitionDef from config.rs to domain::statemachine
where they belong. ports::store now imports only from domain/, fixing
the hexagonal architecture dependency rule violation.

Closes #170
